[Update] Make app can offline

+++ app/src/main/java/vn/asiantech/android/springfinalweather/kotlin/model/InformationWeatherRecyclerView.kt
@@ -7,8 +7,8 @@ class InformationWeatherRecyclerView(val data: List<InformationDetail>)
 
 class InformationDetail(val datetime: String,
                         val weather: WeatherDetail,
-                        @SerializedName("max_temp") val maxTemp: Double,
-                        @SerializedName("min_temp") val minTemp: Double
+                        @SerializedName("max_temp") val maxTemp: Float,
+                        @SerializedName("min_temp") val minTemp: Float
 )
 
 class WeatherDetail(val icon: String,
